+++ app/lib/ui/components/components.dart
@@ -20,6 +20,7 @@ import 'package:ouds_core/components/chip/ouds_filter_chip.dart';
 import 'package:ouds_core/components/divider/ouds_divider.dart';
 import 'package:ouds_core/components/radio_button/ouds_radio_button.dart';
 import 'package:ouds_core/components/switch/ouds_switch.dart';
+import 'package:ouds_core/components/tag/ouds_tag.dart';
 import 'package:ouds_flutter_demo/l10n/app_localizations.dart';
 import 'package:ouds_flutter_demo/ui/components/badge/badge_demo_screen.dart';
 import 'package:ouds_flutter_demo/ui/components/button/button_demo_screen.dart';
@@ -35,6 +36,8 @@ import 'package:ouds_flutter_demo/ui/components/radio_button/radio_button_item_d
 import 'package:ouds_flutter_demo/ui/components/switch/switch_demo_screen.dart';
 import 'package:ouds_flutter_demo/ui/components/switch/switch_item_demo_screen.dart';
 import 'package:ouds_theme_contract/ouds_theme.dart';
+import 'package:ouds_flutter_demo/ui/components/tag/tag_demo_screen.dart';
+import 'package:ouds_flutter_demo/ui/components/tag/tag_input_demo_screen.dart';
 
 List<Component> components(BuildContext context) {
   final theme = OudsTheme.of(context);
@@ -208,6 +211,23 @@ List<Component> components(BuildContext context) {
         VariantComponent(
           context.l10n.app_components_switch_switchItem_label,
           SwitchButtonItemDemoScreen(),
+        ),
+      ],
+    ),
+    Component.withVariant(
+      context.l10n.app_components_tag_label,
+      ComponentContainer(
+        child: OudsTag(label: "Label",status: OudsTagStatus.positive)
+      ),
+      context.l10n.app_components_tag_description_text,
+      [
+        VariantComponent(
+            context.l10n.app_components_tag_label,
+            TagDemoScreen()
+        ),
+        VariantComponent(
+            context.l10n.app_components_tagInput_label,
+            TagInputDemoScreen()
         ),
       ],
     ),