+++ app/lib/ui/theme/theme_controller.dart
@@ -11,7 +11,6 @@
  */
 
 import 'package:flutter/material.dart';
-import 'package:ouds_theme_contract/config/ouds_theme_config_model.dart';
 import 'package:ouds_theme_contract/ouds_theme_contract.dart';
 import 'package:ouds_theme_orange/orange_theme.dart';
 import 'package:ouds_theme_sosh/ouds_theme_sosh.dart';
@@ -20,7 +19,7 @@ class ThemeController extends ChangeNotifier with WidgetsBindingObserver {
   ThemeMode _themeMode = ThemeMode.system;
   OudsThemeContract _currentTheme = OrangeTheme();
   bool _onColoredSurface = false;
-  bool _onBorderRadiusState = false;
+  bool _onBorderRadiusTagState = true;
 
   /// Getter to access the current theme
   OudsThemeContract get currentTheme => _currentTheme;
@@ -32,7 +31,7 @@ class ThemeController extends ChangeNotifier with WidgetsBindingObserver {
   bool get onColoredSurface => _onColoredSurface;
 
   /// Getter to know if we are on a border state
-  bool get onBorderRadiusState => _onBorderRadiusState;
+  bool get onBorderRadiusTagState => _onBorderRadiusTagState;
 
   ThemeController() {
     /// Initialize the theme based on the system's current brightness setting
@@ -93,10 +92,10 @@ class ThemeController extends ChangeNotifier with WidgetsBindingObserver {
     }
   }
 
-  void setOnBorderRadiusState(bool? value) {
+  void setOnBorderRadiusTagState(bool? value) {
     bool newValue = value ?? false;
-    if (_onBorderRadiusState != newValue) {
-      _onBorderRadiusState = newValue;
+    if (_onBorderRadiusTagState != newValue) {
+      _onBorderRadiusTagState = newValue;
       notifyListeners();
     }
   }